+++ storage_class/storage_class.go
@@ -37,11 +37,11 @@ func NewFromPersistent(persistent *StorageClassPersistent) *StorageClass {
 	return New(persistent.Config)
 }
 
-func (s *StorageClass) Matches(vc *storage.StoragePool) bool {
+func (s *StorageClass) Matches(storagePool *storage.StoragePool) bool {
 	if len(s.config.BackendStoragePools) > 0 {
-		if vcList, ok := s.config.BackendStoragePools[vc.Backend.Name]; ok {
-			for _, vcName := range vcList {
-				if vcName == vc.Name {
+		if storagePoolList, ok := s.config.BackendStoragePools[storagePool.Backend.Name]; ok {
+			for _, storagePoolName := range storagePoolList {
+				if storagePoolName == storagePool.Name {
 					return true
 				}
 			}
@@ -49,19 +49,19 @@ func (s *StorageClass) Matches(vc *storage.StoragePool) bool {
 	}
 	matches := len(s.config.Attributes) > 0
 	for name, request := range s.config.Attributes {
-		if vc.Attributes == nil {
+		if storagePool.Attributes == nil {
 			log.WithFields(log.Fields{
 				"storageClass": s.GetName(),
-				"pool":         vc.Name,
+				"pool":         storagePool.Name,
 				"attribute":    name,
 			}).Panic("Storage pool attributes are nil")
 		}
-		if offer, ok := vc.Attributes[name]; !ok || !offer.Matches(request) {
+		if offer, ok := storagePool.Attributes[name]; !ok || !offer.Matches(request) {
 			log.WithFields(log.Fields{
 				"offer":        offer,
 				"request":      request,
 				"storageClass": s.GetName(),
-				"pool":         vc.Name,
+				"pool":         storagePool.Name,
 				"attribute":    name,
 				"found":        ok}).Debug("Attribute for storage " +
 				"pool failed to match storage class.")
@@ -80,10 +80,10 @@ func (s *StorageClass) CheckAndAddBackend(b *storage.StorageBackend) int {
 		return 0
 	}
 	added := 0
-	for _, vc := range b.Storage {
-		if s.Matches(vc) {
-			s.pools = append(s.pools, vc)
-			vc.AddStorageClass(s.GetName())
+	for _, storagePool := range b.Storage {
+		if s.Matches(storagePool) {
+			s.pools = append(s.pools, storagePool)
+			storagePool.AddStorageClass(s.GetName())
 			added++
 		}
 	}
@@ -91,19 +91,19 @@ func (s *StorageClass) CheckAndAddBackend(b *storage.StorageBackend) int {
 }
 
 func (s *StorageClass) RemovePoolsForBackend(backend *storage.StorageBackend) {
-	newVCList := make([]*storage.StoragePool, 0)
-	for _, vc := range s.pools {
-		if vc.Backend != backend {
-			newVCList = append(newVCList, vc)
+	newStoragePools := make([]*storage.StoragePool, 0)
+	for _, storagePool := range s.pools {
+		if storagePool.Backend != backend {
+			newStoragePools = append(newStoragePools, storagePool)
 		}
 	}
-	s.pools = newVCList
+	s.pools = newStoragePools
 }
 
 func (s *StorageClass) GetVolumes() []*storage.Volume {
 	ret := make([]*storage.Volume, 0)
-	for _, vc := range s.pools {
-		for _, vol := range vc.Volumes {
+	for _, storagePool := range s.pools {
+		for _, vol := range storagePool.Volumes {
 			// Because storage pools can fulfill more than one storage
 			// class, we have to check each volume for that pool to see
 			// if it uses this storage class.
@@ -130,9 +130,9 @@ func (s *StorageClass) GetBackendStoragePools() map[string][]string {
 func (s *StorageClass) GetStoragePoolsForProtocol(p config.Protocol) []*storage.StoragePool {
 	ret := make([]*storage.StoragePool, 0, len(s.pools))
 	// TODO:  Change this to work with indices of backends?
-	for _, vc := range s.pools {
-		if p == config.ProtocolAny || vc.Backend.GetProtocol() == p {
-			ret = append(ret, vc)
+	for _, storagePool := range s.pools {
+		if p == config.ProtocolAny || storagePool.Backend.GetProtocol() == p {
+			ret = append(ret, storagePool)
 		}
 	}
 	return ret
@@ -143,25 +143,25 @@ func (s *StorageClass) ConstructExternal() *StorageClassExternal {
 		Config:       s.config,
 		StoragePools: make(map[string][]string),
 	}
-	for _, vc := range s.pools {
-		backendName := vc.Backend.Name
-		if vcList, ok := ret.StoragePools[backendName]; ok {
+	for _, storagePool := range s.pools {
+		backendName := storagePool.Backend.Name
+		if storagePoolList, ok := ret.StoragePools[backendName]; ok {
 			log.WithFields(log.Fields{
 				"storageClass": s.GetName(),
-				"pool":         vc.Name,
+				"pool":         storagePool.Name,
 				"Backend":      backendName,
 				"Method":       "ConstructExternal",
 			}).Debug("Appending to existing storage pool list for backend.")
-			ret.StoragePools[backendName] = append(vcList, vc.Name)
+			ret.StoragePools[backendName] = append(storagePoolList, storagePool.Name)
 		} else {
 			log.WithFields(log.Fields{
 				"storageClass": s.GetName(),
-				"pool":         vc.Name,
+				"pool":         storagePool.Name,
 				"Backend":      backendName,
 				"Method":       "ConstructExternal",
 			}).Debug("Creating new storage pool list for backend.")
 			ret.StoragePools[backendName] = make([]string, 1, 1)
-			ret.StoragePools[backendName][0] = vc.Name
+			ret.StoragePools[backendName][0] = storagePool.Name
 		}
 	}
 	for _, list := range ret.StoragePools {